user:schmidst/stsch47
Bugfix AKSEN-Protokoll
Update aksentest

git-svn-id: http://svn.code.sf.net/p/segelboot/code@241 2e2d5edc-86b9-48e7-b817-bf6befc431f3

+++ trunk/src/de/fhb/sailboat/serial/actuator/LocomotionSystem.java
@@ -62,6 +62,6 @@ public interface LocomotionSystem {
 	public void resetPropellor();
 	public int getBatteryState();
 	public void setDebug(boolean debug);
-	public int getWait_Sleep();
+	public long getWait_Sleep();
 	public void setWait_Sleep(int wait_sleep);
 }
